requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pillow>=10.0.0
numpy>=1.25.0
sympy>=1.13.0
